Improved the User booking

+++ api/venueApi.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, fields, marshal_with, reqparse
-from models.admin import Venue
+from models.admin import Venue, Allocation
 from db import db
 from validation import NotFoundError, BusinessValidationError
 from sqlalchemy import desc
@@ -11,7 +11,7 @@ venue_output_fields = {
     "city" : fields.String,
     "capacity" : fields.Integer,
     "description" : fields.String,
-    "timestamp" : fields.String
+    "timestamp" : fields.DateTime(dt_format='rfc822')
 
 }
 
@@ -41,34 +41,34 @@ class VenueAPI(Resource):
         city = vn_args.get('city',None)
         capacity = vn_args.get('capacity',None)
         desc = vn_args.get('description',None)
-        from datetime import datetime
-        timestamp = datetime.now()
          
 
-        if name is None or name is '':
+        if name is None or name == '':
             raise BusinessValidationError(status_code=400,error_code="VN002",error_message="Name is required")
     
-        if location is None or location is '':
+        if location is None or location == '':
             raise BusinessValidationError(status_code=400,error_code="VN003",error_message="Location is required")
 
-        if capacity is None or capacity is '':
-            raise BusinessValidationError(status_code=400,error_code="VN004",error_message="Capacity is required")
+        if city is None or city == '':
+            raise BusinessValidationError(status_code=400,error_code="VN004",error_message="City is required")
 
-        if int(capacity) <40:
+        if capacity == None or type(capacity) is not int:
             raise BusinessValidationError(status_code=400,error_code="VN005",error_message="Invalid capacity of venue")
         
-        if desc is None or desc is '':
-            raise BusinessValidationError(status_code=400,error_code="VN006",error_message="Description is required")
-    
-        if city is None or city is '':
-            raise BusinessValidationError(status_code=400,error_code="VN007",error_message="City is required")
-    
+        if int(capacity) < 40:
+            raise BusinessValidationError(status_code=400,error_code="VN006",error_message="Capacity must be atleast 40")
+        
+        if desc is None or desc == '':
+            raise BusinessValidationError(status_code=400,error_code="VN007",error_message="Description is required")
 
 
         venue = db.session.query(Venue).filter(Venue.name == name).first()
 
         if venue:
-            raise BusinessValidationError(status_code=400,error_code="VN007",error_message="Venue already exists")
+            raise BusinessValidationError(status_code=400,error_code="VN008",error_message="Venue already exists")
+
+        from datetime import datetime
+        timestamp = datetime.now()
 
         new_venue = Venue(name=name,location=location,city=city,capacity=int(capacity),description=desc,timestamp=timestamp)
         db.session.add(new_venue)
@@ -85,10 +85,14 @@ class VenueAPI(Resource):
             raise BusinessValidationError(status_code=400,error_code="VN011",error_message="Venue not found with such name")
         
         # check for dependency
+        showForVenue = db.session.query(Allocation).filter(Allocation.venue_id == venue.id).first()
+
+        if showForVenue:
+            raise BusinessValidationError(status_code=400,error_code="VN012",error_message="Venue has shows allocated to it")
 
         db.session.delete(venue)
         db.session.commit()
-        return "Succesfully Deleted", 200
+        return "Success", 200
 
     
 class VenueListByCityApi(Resource):
@@ -100,7 +104,7 @@ class VenueListByCityApi(Resource):
         if venues:
             return venues
         else:
-            raise NotFoundError(error_message='No Venues found for this city',status_code=404,error_code="VN008")
+            raise NotFoundError(error_message='No Venues found for this city',status_code=404,error_code="VN009")
         
     def post(self):
         pass
@@ -109,7 +113,7 @@ class VenueListByNameApi(Resource):
 
     @marshal_with(venue_output_fields)
     def get(self,name):
-        venues = db.session.query(Venue).filter(Venue.name.ilike(f'%{name}%')).all()
+        venues = db.session.query(Venue).filter(Venue.name.ilike(f'%{name}%')).order_by(desc(Venue.timestamp)).all()
 
         if venues:
             return venues
